% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{accept}
\alias{accept}
\title{chainable function to access add_accepts method}
\usage{
accept(bp, x, overwrite = FALSE)
}
\arguments{
\item{bp}{blueprint object}

\item{x}{vector of accept strings}

\item{overwrite}{whether to overwrite existing accept strings, default FALSE}
}
\description{
chainable function to access add_accepts method
}
\examples{
bp <- Blueprint$new("nonmem")
bp \%>\%
accept(c("TIME.LT.100", "DV.GT.5"))
}
