% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-manip.R
\name{append_classes}
\alias{append_classes}
\title{append classes to a set of object classes}
\usage{
append_classes(.x, new_classes)
}
\arguments{
\item{.x}{object}

\item{new_classes}{character string or vector of new classes to append}
}
\description{
append classes to a set of object classes
}
\keyword{internal}
