% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.as.factor.R
\name{attrs.as.factor}
\alias{attrs.as.factor}
\title{attrs.as.factor}
\usage{
attrs.as.factor(x)
}
\arguments{
\item{x}{a vector object or data.fram.}
}
\value{
Either a factor (if x is a vector) containing the attributes as levels or a data.frame with every column replaced by according factors.
}
\description{
Run over attributes and try to create a vector.
}
\author{
Marc Schwenzer <m.schwenzer@uni-tuebingen.de>
}

