% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block}
\alias{block}
\title{block matrix}
\usage{
block(..., param_names, fixed = FALSE, correlation = FALSE,
  comment = NULL)
}
\arguments{
\item{...}{elements to coerce into a block matrix}

\item{param_names}{parameters associated with the elements of the matrix}

\item{fixed}{whether the values should be fixed}

\item{correlation}{whether to create as a correlation matrix}

\item{comment}{comment to append}
}
\description{
block matrix
}
\examples{
block(0.04, 0.01, 0.09, param_names = c("CL", "V"))
# can also set as correlation matrix
block(0.04, 0.61, 0.09, param_names = c("CL", "V"), correlation = TRUE)
}
