% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_info.R
\name{blue_info}
\alias{blue_info}
\title{blue_info}
\usage{
blue_info(blueprint = options("blueprint_file"), waves, searchstr = NULL)
}
\arguments{
\item{blueprint}{The Blueprint file.}

\item{waves}{The wave to inform about. By now only numeric vector of length 1 accepted.}

\item{search}{An optional pattern to search for in the labels.}
}
\value{
Information
}
\description{
\code{blue_example}
}
\details{
Inform about a blueprint
}
\author{
Marc Schwenzer <m.schwenzer@uni-tuebingen.de>
}

