% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_set.R
\name{blue_set}
\alias{blue_set}
\title{blue_set}
\usage{
blue_set(blueprint)
}
\arguments{
\item{blueprint}{}
}
\description{
blue_set()
}
\details{
Easily set a default blueprint file. This is taken into account by blue(), blue_open(), blue_info() and alike functions.
}
\author{
Marc Schwenzer
}

