% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-package.R
\docType{package}
\name{blueprint-package}
\alias{blueprint}
\alias{blueprint-package}
\title{Blueprint: an efficient data schemas generator for R}
\description{
An unopinionated consistent interface to generate 
    all sorts of data schemas in R. Data schemas are R objects
    that dictate what should go in R data structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blueprint.potvin.xyz}
  \item \url{https://github.com/jeanmathieupotvin/blueprint}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/blueprint/issues}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jm@potvin.xyz} (\href{https://orcid.org/0000-0002-8237-422X}{ORCID})

}
\keyword{internal}
