% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluepRint.R
\name{blueprint}
\alias{blueprint}
\title{Layout function}
\usage{
blueprint(layout, ..., label = TRUE)
}
\description{
This function provides an interface to plot common layouts for ggplot2
}
\examples{
blueprint('square', plot1, plot2, plot3, plot4)
blueprint('square', plot1, plot2, plot3, plot4, label=F)
}

