% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprinter.R
\name{\%bp\%}
\alias{\%bp\%}
\alias{Blueprinter}
\title{Concise interface to blueprint}
\usage{
symbol \%bp\% object
}
\arguments{
\item{symbol}{A syntactic \link{name}, also known as an \emph{unquoted name}.}

\item{object}{Any \R value.}
}
\value{
The \code{\link[=Blueprinter]{\%bp\%}} operator is a convenient wrapper to
\pkg{blueprint}'s constructors and returns an appropriate \link[R6:R6Class]{R6}
object, a \link{Blueprint} that corresponds to the underlying type of \code{value}.
\itemize{
\item If value is of a \emph{strict atomic} type (one of \link[base:NULL]{NULL},
\link[base:logical]{logical}, \link[base:integer]{integer},
\link[base:numeric]{numeric}, \link[base:complex]{complex},
\link[base:character]{character} or \link[base:raw]{raw}), an
instance of class \link{Atomic} is returned.
}
}
\description{
Create instances of \pkg{blueprint}'s classes faster and more easily.
}
