% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\name{const}
\alias{const}
\title{create a constant parameter}
\usage{
const(.value, .comment)
}
\arguments{
\item{.value}{value to hold}

\item{.comment}{comment}
}
\description{
create a constant parameter
}
