% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{constants}
\alias{constants}
\title{chainable method to add_constants}
\usage{
constants(bp, ...)
}
\arguments{
\item{bp}{blueprint object}

\item{...}{params to pass}
}
\description{
chainable method to add_constants
}
