% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{covariate_relationships}
\alias{covariate_relationships}
\alias{covariate_relationships.parameter}
\title{get covariate relationships for a parameter}
\usage{
covariate_relationships(x, ...)

\method{covariate_relationships}{parameter}(x)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\examples{
cl <- parameter(1.5, covariate_relationships = list(WT = ("WT/70")))
covariate_relationships(cl)
}
