% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des.atomic}
\alias{des.atomic}
\title{des.atomic}
\usage{
des.atomic(x)
}
\arguments{
\item{x}{}
}
\value{
a data.frame
}
\description{
des.atomic
}
\details{
Wrapper to deal with atomic values by encapsulating values into a data.frame.
}
\author{
Marc Schwenzer
}
