% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des.data.frame}
\alias{des.data.frame}
\title{des.data.frame}
\usage{
des.data.frame(x, hist = "", tab)
}
\arguments{
\item{x}{}

\item{hist}{}

\item{tab}{}
}
\value{
data.frame with statistics in column.
}
\description{
des.data.frame
}
\details{
des.data.frame
}
\author{
Marc Schwenzer
}
