% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des}
\alias{des}
\title{des}
\usage{
des(df, ..., hist = FALSE, tab = FALSE)
}
\arguments{
\item{df}{Either a single variable or a data.frame for which descriptive sumamry statistics will be generated. If df contains a variable named `weight` this will be used to weight the data. If df is of class `grouped_df` the descripitve statistics will be generated for every group.}

\item{...}{}

\item{hist}{}

\item{tab}{}
}
\value{
A data.frame of class `tbl_df`.
}
\description{
des
}
\details{
Summarise certain variables by group.
}
\author{
Marc Schwenzer
}
