% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{fixed}
\alias{fixed}
\alias{fixed.parameter}
\title{get whether a parameter is fixed}
\usage{
fixed(x, ...)

\method{fixed}{parameter}(x)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\examples{
cl <- parameter(1.5, fixed = TRUE)
fixed(cl)
cl <- update(cl, fixed = FALSE)
fixed(cl)
}
