% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fre.R
\name{fre}
\alias{fre}
\title{fre}
\usage{
fre(x, groupingvars = NULL, maxcat = 25, print = TRUE)
}
\arguments{
\item{x}{A vector or data.frame. If x is a data frame, it is iterated over its columns.}

\item{maxcat}{Maximum number of categories to be printed. Default: 50}
}
\value{
data.frame (invisible)
}
\description{
fre
}
\details{
Compute a simple frequency summary and print it.
}
\author{
Marc Schwenzer
}
