% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{from_path}
\alias{from_path}
\title{chainable method to set path to data file}
\usage{
from_path(bp, path)
}
\arguments{
\item{bp}{blueprint object}

\item{path}{path for data file for a control stream}
}
\description{
chainable method to set path to data file
}
