% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{hierarchies}
\alias{hierarchies}
\title{chainable method to add hierarchies}
\usage{
hierarchies(bp, ...)
}
\arguments{
\item{bp}{blueprint object}

\item{...}{params to pass}
}
\description{
chainable method to add hierarchies
}
