% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{hooks}
\alias{hooks}
\title{chainable method to add hooks}
\usage{
hooks(bp, ...)
}
\arguments{
\item{bp}{blueprint object}

\item{...}{params to pass}
}
\description{
chainable method to add hooks
}
\details{
hooks can be added into any template, and should
follow the naming convention hooks:<major_element>:<location>

For example, {{hooks:pk:begin}} might be
placed at the top of $PK in a template

the  [hooks()] function will inject a string into that template
by matching the name.
}
\examples{
library(dplyr)
bp <- Blueprint$new("nonmem")

bp$template <- "
$PK
{{hooks:pk:begin}}

// stuff in $PK

{{hooks:pk:end}}
"
bp \%>\%
hooks("pk:begin" = "// comment at top of PK",
      "pk:end" = "V2 = S2/1000") \%>\%
render() \%>\%
// to print lines
cat()
}
