% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{ignore}
\alias{ignore}
\title{chainable function to access add_ignores method}
\usage{
ignore(bp, x, overwrite = FALSE)
}
\arguments{
\item{bp}{blueprint object}

\item{x}{vector of ignore strings}
}
\description{
chainable function to access add_ignores method
}
\examples{
bp <- Blueprint$new("nonmem")
bp \%>\%
ignore(c("TIME.GT.100", "DV.LT.1"))
}
