% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_design.R
\name{is_trial_design}
\alias{is_trial_design}
\title{Test if an object is a trial design}
\usage{
is_trial_design(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{trial_design}, else \code{FALSE}.
}
\description{
Test if an object is a trial design
}
\examples{
design <- design_simon(n1 = 13, r1 = 3, n = 43, r = 12, p0 = 0.2, pa = 0.4)
is_trial_design(design)

}
