% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{link}
\alias{link}
\alias{link.parameter}
\title{get a parameters link to a heirarchy}
\usage{
link(x, ...)

\method{link}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\description{
get the link for a parameter
}
\examples{
cl <- parameter(1.5, name = "CL")
link(cl)
cl2 <- parameter(1.5, name = "TVCL", link = "CL")
link(cl2)
}
