% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-partials.R
\name{load_partials}
\alias{load_partials}
\title{load partials from a given type}
\usage{
load_partials(type)
}
\arguments{
\item{type}{type - nonmem or mrgsolve}
}
\description{
load partials from a given type
}
