% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{lower_bound}
\alias{lower_bound}
\alias{lower_bound.parameter}
\title{get a parameters lower bound}
\usage{
lower_bound(x, ...)

\method{lower_bound}{parameter}(x)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\examples{
cl <- parameter(1.5)
lower_bound(cl)
cl <- update(cl, lower_bound = 0)
lower_bound(cl)
}
