% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{model_type}
\alias{model_type}
\title{chainable method to specify the model type}
\usage{
model_type(bp, model_type)
}
\arguments{
\item{bp}{blueprint object}

\item{model_type}{type of model}
}
\description{
chainable method to specify the model type
}
\examples{
bp <- Blueprint$new("nonmem")
models <- available_models("nonmem")
bp \%>\% model_type(models$one_cmt_iv)
}
