% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{name}
\alias{name}
\alias{name.parameter}
\title{get a parameters name}
\usage{
name(x, ...)

\method{name}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\description{
get the name of the parameter
}
\examples{
cl <- parameter(1.5, name = "TVCL")
name(cl)
}
