% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{names.block}
\alias{names.block}
\alias{names.sigma}
\title{get the name of the parameter}
\usage{
\method{names}{block}(x, ...)

\method{names}{sigma}(x, ...)
}
\description{
get the name of the parameter

get the name of the parameter
}
