% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{numeric.descriptives}
\alias{numeric.descriptives}
\title{numeric.descriptives}
\usage{
numeric.descriptives(df)
}
\arguments{
\item{df}{}
}
\value{
a data.frame row of numeric descriptives
}
\description{
numeric.descriptives
}
\details{
compute numeric descriptives. Acutally workhorse behind des.
}
\author{
Marc Schwenzer
}
