% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.R
\name{open_blue}
\alias{open_blue}
\title{open_blue creates or loads blueprint-file .}
\usage{
open_blue(blueprint = paste0(getwd(), "/blueprint.xlsx"), waves = 1,
  debug = FALSE)
}
\arguments{
\item{blueprint}{Path to blueprint (meta-data file) that contains specifications about the variables in data files that will be merged. The file format of data is taken from the suffix as defined in the \code{\link[=rio]{import}}. See the vignette for details of the structure.}

\item{waves}{Number specifying how many waves shall be included in the new blueprint file. This can be changed later manually by adding appropriate named columns.}
}
\value{
Returns nothing. It is just used for the side effect of generating or opening the specified blueprint file.
}
\description{
\code{"open_blue"} creates or loads blueprint-files of various file formats.
}
\examples{
open_blue('/path/to/file.xlsx')
}
\author{
Marc Schwenzer <m.schwenzer@uni-tuebingen.de>
}

