% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\name{param}
\alias{param}
\title{create details about a parameter}
\usage{
param(.value, .comment = NULL, .lower_bound = NULL, .upper_bound = NULL,
  .fixed = FALSE, covariate_relationships = NULL)
}
\arguments{
\item{.value}{initial value of a parameter}

\item{.lower_bound}{a lower bound}

\item{.upper_bound}{an upper bound}

\item{.fixed}{whether the parameter should be fixed}

\item{covariate_relationships}{covariate relationships to a parameter}
}
\description{
create details about a parameter
}
