% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\docType{data}
\name{Parameter}
\alias{Parameter}
\title{Parameter class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Parameter
}
\description{
Parameter class
}
\keyword{datasets}
