% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{parameters}
\alias{parameters}
\title{chainable function to access add_params method}
\usage{
parameters(bp, ...)
}
\arguments{
\item{bp}{blueprint object}

\item{...}{params to pass}
}
\description{
chainable function to access add_params method
}
