% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-manip.R
\name{prepend_classes}
\alias{prepend_classes}
\title{prepend classes to a set of object classes}
\usage{
prepend_classes(.x, new_classes)
}
\arguments{
\item{.x}{object}

\item{new_classes}{character string or vector of new classes to prepend}
}
\description{
prepend classes to a set of object classes
}
\keyword{internal}
