% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.basket_trial_design}
\alias{print.basket_trial_design}
\title{Print method for basket_trial_design results}
\usage{
\method{print}{basket_trial_design}(x, ...)
}
\arguments{
\item{x}{Output from \code{two_stage_basket_trial}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for basket_trial_design results
}
