% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.simple_basket_design}
\alias{print.simple_basket_design}
\title{Print method for simple_basket_design results}
\usage{
\method{print}{simple_basket_design}(x, ...)
}
\arguments{
\item{x}{Output from \code{simon_basket_trial}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for simple_basket_design results
}
