% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.basket_trial_design}
\alias{print.summary.basket_trial_design}
\title{Print summary for basket trial design}
\usage{
\method{print}{summary.basket_trial_design}(x, ...)
}
\arguments{
\item{x}{Summary object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print summary for basket trial design
}
