% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.simple_basket_design}
\alias{print.summary.simple_basket_design}
\title{Print summary for Simon basket design}
\usage{
\method{print}{summary.simple_basket_design}(x, ...)
}
\arguments{
\item{x}{Summary object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print summary for Simon basket design
}
