% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.two_stage_design}
\alias{print.summary.two_stage_design}
\title{Print summary for two-stage design}
\usage{
\method{print}{summary.two_stage_design}(x, ...)
}
\arguments{
\item{x}{Summary object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print summary for two-stage design
}
