% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.two_stage_design}
\alias{print.two_stage_design}
\title{Print method for two_stage_single_arm_tte results}
\usage{
\method{print}{two_stage_design}(x, ...)
}
\arguments{
\item{x}{Output from \code{two_stage_single_arm_tte}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for two_stage_single_arm_tte results
}
