% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{residual_error}
\alias{residual_error}
\title{chainable method to add_residual_error}
\usage{
residual_error(bp, ...)
}
\arguments{
\item{bp}{blueprint object}

\item{...}{params to pass}
}
\description{
chainable method to add_residual_error
}
