% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_spl.R
\name{select_spl}
\alias{select_spl}
\title{select_spl}
\usage{
select_spl(df, .dots, odf = "odf")
}
\arguments{
\item{df}{A data.frame to split}

\item{.dots}{A string of variables to select (which has to be conform to inpute epected by select_).}

\item{odf}{Name of a newly created other df with the variables not matching. Defaulting to `odf` short for other data.frame.}
}
\value{
The data.frame with matchin columns.
}
\description{
Help function to split a df into two disjunct dfs
}
\details{
select_spl splits a data.frame into two, one containing some specified variables and one not containing some variables. Like dplyr::select it returns the df with matching columns but additionally writes the other non-matching columns to a object in the calling environment. This is a potentially destructive function for its sideeffect of creating other.df.
}
\author{
Marc Schwenzer
}
