% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_blue.R
\name{set_blue}
\alias{set_blue}
\title{set_blue}
\usage{
set_blue(blueprint)
}
\arguments{
\item{blueprint}{}
}
\value{
NA
}
\description{
set_blue()
}
\details{
Easily set a default blueprint file. This is taken into account by blue(), blue_open(), blue_info() and alike functions.
}
\author{
Marc Schwenzer
}
