% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-operations.R
\name{stringify_matrix}
\alias{stringify_matrix}
\title{return the scaled covariance matrix as a string to embed in a file, in scientific notation}
\usage{
stringify_matrix(.mat, .fixed = FALSE)
}
\arguments{
\item{.mat}{matrix}

\item{.fixed}{whether to include FIXED in the first line (for nonmem)}
}
\description{
return the scaled covariance matrix as a string to embed in a file, in scientific notation
}
