% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.basket_trial_design}
\alias{summary.basket_trial_design}
\title{Summary method for basket_trial_design results}
\usage{
\method{summary}{basket_trial_design}(object, ...)
}
\arguments{
\item{object}{Output from \code{two_stage_basket_trial}}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summary method for basket_trial_design results
}
