% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.simple_basket_design}
\alias{summary.simple_basket_design}
\title{Summary method for simple_basket_design results}
\usage{
\method{summary}{simple_basket_design}(object, ...)
}
\arguments{
\item{object}{Output from \code{simon_basket_trial}}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summary method for simple_basket_design results
}
