% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.two_stage_design}
\alias{summary.two_stage_design}
\title{Summary method for two_stage_single_arm_tte results}
\usage{
\method{summary}{two_stage_design}(object, ...)
}
\arguments{
\item{object}{Output from \code{two_stage_single_arm_tte}}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summary method for two_stage_single_arm_tte results
}
