% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{upper_bound}
\alias{upper_bound}
\alias{upper_bound.parameter}
\title{get a parameters upper bound}
\usage{
upper_bound(x, ...)

\method{upper_bound}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\description{
get the upper bound for a parameter
}
\examples{
cl <- parameter(1.5, lower_bound = 0)
upper_bound(cl)
cl <- update(cl, upper_bound = 100)
upper_bound(cl)
}
