% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{use_template}
\alias{use_template}
\title{chainable method to set template}
\usage{
use_template(bp, template)
}
\arguments{
\item{bp}{blueprint object}

\item{template}{template to use}
}
\description{
chainable method to set template
}
