% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_var.R
\name{validate_var}
\alias{validate_var}
\title{validate_var}
\usage{
validate_var(data, var, min = NA, max = NA)
}
\arguments{
\item{data}{A data.frame}

\item{var}{The variable to analyse}

\item{min}{Lower limit, values below are counted}

\item{max}{Upper limit, values above are counted}
}
\description{
Compute statistics over a variable in a data.frame.
}
\details{
Works also with grouped_df
}
\author{
Marc Schwenzer
}
