% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{value}
\alias{value}
\alias{value.parameter}
\title{get the value of a parameter}
\usage{
value(x, ...)

\method{value}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\description{
get the value for a parameter
}
\examples{
cl <- parameter(1.5)
value(cl)
}
