% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{with_data}
\alias{with_data}
\title{chainable method to set the data}
\usage{
with_data(bp, data)
}
\arguments{
\item{bp}{blueprint object}

\item{data}{dataset}
}
\description{
chainable method to set the data
}
