% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.R
\name{blue_example}
\alias{blue_example}
\title{Creates a folder with example files

This folder will be named 'blueprint_example' and contains the example files 'INT_STU12_DEC03_synth.sav','INT_SCQ12_DEC03_synth.sav','example_blueprint1.xlsx','example_blueprint1.csv','example_blueprint2.xlsx','example_blueprint2.csv'.}
\usage{
blue_example()
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Marc Schwenzer
}

