% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\docType{data}
\name{Blueprint}
\alias{Blueprint}
\title{base blueprint class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Blueprint
}
\description{
base blueprint class
}
\keyword{datasets}
