% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\docType{package}
\name{blueprint}
\alias{blueprint}
\alias{blueprint-package}
\title{blueprint: a tool to auto-merge datasets by rules specified in a meta-file}
\description{
blueprint offers the opportunity to
\itemize{\item{select}
\item{combine}
\item{transform}
} variables (columns) from various data file. It establishs an easy and intutitive way to merge data by using a meta-data file (a "blueprint"). This file can be easily edited with a choosen spreadsheet application or text editor. Possible formats in which the meta-data may be contained are: CSV, Excel, OpenOffice Calc, HTML, ....
}
\details{
Have a look at the vignette:

\code{browseVignettes(package = "blueprint")}
}

