% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprinter.R
\name{\%bp\%}
\alias{\%bp\%}
\alias{Blueprinter}
\title{Concise interface to blueprint}
\usage{
symbol \%bp\% object
}
\arguments{
\item{symbol}{A syntactic \link{name}, also known as an \emph{unquoted name}, or a
scalar character.}

\item{object}{Any \R value.}
}
\value{
The \code{\link[=Blueprinter]{\%bp\%}} operator is a convenient wrapper to
\pkg{blueprint}'s constructor functions and always returns an appropriate
\link[R6:R6Class]{R6} object. The output is always a \link{Blueprint} that corresponds
to the underlying type of \code{object}.
\itemize{
\item If \code{object} is a \link[=is_strict_atomic]{strict atomic value},
an \link{Atomic} object is returned.
}

If no \link{Blueprint} class corresponding to \code{object} is available, a
\code{NULL} value is returned with a warning.
}
\description{
Create instances of \pkg{blueprint}'s classes with a fast and concise
operator. The output's class is automatically deduced from argument
\code{object}.
}
