% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des}
\alias{des}
\title{des}
\usage{
des(df, ..., hist = FALSE, tab = FALSE)
}
\arguments{
\item{df}{}

\item{...}{}

\item{g}{}
}
\value{
A tbl
}
\description{
des
}
\details{
Summarise certain variables by group
}
\author{
Marc Schwenzer
}
