% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_error_eqn.R
\name{get_residual_error_eqn}
\alias{get_residual_error_eqn}
\title{get the residual error equation}
\usage{
get_residual_error_eqn(elements, ipred, type)
}
\arguments{
\item{elements}{residual error elements}

\item{ipred}{the name for the IPRED variable}

\item{type}{type of model, nonmem or mrgsolve}
}
\description{
get the residual error equation
}
\details{
currently supports ADD and PROP
}
\examples{
get_residual_error_eqn("ADD", "IPRED", "nonmem")
get_residual_error_eqn(c("ADD", "PROP"), "IPRED", "nonmem")
get_residual_error_eqn(c("PROP"), "CP", "mrgsolve")
}
