% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_blue.R
\name{info_blue}
\alias{info_blue}
\title{info_blue}
\usage{
info_blue(blueprint = options("blueprint_file"), waves, searchstr = NULL)
}
\arguments{
\item{blueprint}{The Blueprint file.}

\item{waves}{The wave to inform about. By now only numeric vector of length 1 accepted.}

\item{search}{An optional pattern to search for in the labels.}
}
\value{
Information
}
\description{
\code{info_blue}
}
\details{
Inform about a blueprint
}
\author{
Marc Schwenzer <m.schwenzer@uni-tuebingen.de>
}

