% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{omega_param}
\alias{omega_param}
\title{diagonal values}
\usage{
omega_param(value, link, fix = FALSE, comment = NULL)
}
\arguments{
\item{value}{value}

\item{link}{link parameter name}

\item{fix}{whether fixed DEFAULT: FALSE}

\item{comment}{comment}
}
\description{
diagonal values
}
