% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\name{param}
\alias{param}
\title{create details about a parameter}
\usage{
param(.value, .bounds = NULL, .fixed = FALSE,
  covariate_relationships = NULL)
}
\arguments{
\item{.value}{initial value of a parameter}

\item{.bounds}{a vector of lower and upper bounds}

\item{.fixed}{whether the parameter should be fixed}

\item{covariate_relationships}{covariate relationships to a parameter}
}
\description{
create details about a parameter
}
