% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R, R/parameter_details.R
\name{new_blueprint}
\alias{new_blueprint}
\alias{parameter}
\title{create a blueprint object}
\usage{
new_blueprint(.type = NULL)

parameter(value, name = NULL, comment = NULL, lower_bound = -Inf,
  upper_bound = Inf, fixed = FALSE, covariate_relationships = NULL,
  link = name)
}
\arguments{
\item{.type}{type of model object to create (eg: nonmem/mrgsolve)}

\item{value}{value of parameter}

\item{name}{name of parameter, Default: NULL}

\item{comment}{comments about the parameter, Default: NULL}

\item{lower_bound}{lower bound, Default: -Inf}

\item{upper_bound}{upper bound, Default: Inf}

\item{fixed}{whether parameter should be fixed, Default: FALSE}

\item{covariate_relationships}{covariate relationships, Default: NULL}

\item{link}{name to link to a hierarchy, Default: name}
}
\value{
object with class parameter

object with class parameter
}
\description{
create a blueprint object

create a parameter
}
\examples{
bp <- new_blueprint("nonmem")

# can just start templating, but will need to set a type before rendering
bp <- new_blueprint()
cl <- parameter(1.5, "CL")
V <- parameter(10.5, "TVV", lower_bound = 0, link = "V")
}
