% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\name{parameter}
\alias{parameter}
\title{create a parameter}
\usage{
parameter(value, name = NULL, comment = NULL, lower_bound = -Inf,
  upper_bound = Inf, fixed = FALSE, covariate_relationships = NULL,
  link = name)
}
\arguments{
\item{value}{value of parameter}

\item{name}{name of parameter, Default: NULL}

\item{comment}{comments about the parameter, Default: NULL}

\item{lower_bound}{lower bound, Default: -Inf}

\item{upper_bound}{upper bound, Default: Inf}

\item{fixed}{whether parameter should be fixed, Default: FALSE}

\item{covariate_relationships}{covariate relationships, Default: NULL}

\item{link}{name to link to a heirarchy, Default: name}
}
\value{
object with class parameter
}
\description{
create a parameter
}
\examples{
cl <- parameter(1.5, "CL")
V <- parameter(10.5, "TVV", lower_bound = 0, link = "V")
}
