% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_parameter.R
\name{parameters.nonmem_blueprint}
\alias{parameters.nonmem_blueprint}
\title{generic setting of parameters for blueprints}
\usage{
parameters.nonmem_blueprint(.bp, ...)
}
\description{
generic setting of parameters for blueprints
}
