% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{sigma_param}
\alias{sigma_param}
\title{diagonal values}
\usage{
sigma_param(.value, .fix = FALSE, .comment = NULL)
}
\arguments{
\item{.value}{value}

\item{.fix}{whether fixed DEFAULT: FALSE}

\item{.comment}{comment}
}
\description{
diagonal values
}
