% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/parameter_details.R
\name{update}
\alias{update}
\alias{update.parameter}
\title{update a parameters information}
\usage{
update(x, ...)

\method{update}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter}

\item{...}{other arguments}
}
\description{
update a parameter
}
\examples{
cl <- parameter(1.5)
cl <- update(cl, lower_bound = 0)
}
